refactor: single <Layout /> for all components
